여러가지 수정(여기 다시 시작)

+++ src/main/java/hello/fifthweek/infrastructure/product/jparepository/ProductIndeAmountJpaRepository.java
@@ -12,11 +12,11 @@ public interface ProductIndeAmountJpaRepository extends JpaRepository<ProductInd
 
 
     //    @Query("SELECT p FROM ProductInfoEntity p JOIN FETCH p.options WHERE p.id = :productId")
-//    @Query(value = "SELECT PRODUCT_ID, PRODUCT_NAME, PRODUCT_OPTION, PRODUCT_QUANTITY FROM PRODUCT_INFO p JOIN FETCH PRODUCT_OPTION_INFO WHERE p.PRODUCT_ID = :productId AND PRODUCT_OPTION_INFO = :productOptionName", nativeQuery = true)
+//    @Query(value = "SELECT PRODUCT_ID, PRODUCT_NAME, PRODUCT_OPTION, PRODUCT_QUANTITY FROM PRODUCT_INFO p JOIN FETCH PRODUCT_OPTION_INFO WHERE p.PRODUCT_ID = :productId AND PRODUCT_OPTION_INFO = :productOptionId", nativeQuery = true)
 //    ProductIndeAmountInfraEntity findByProductId(@Param("productId") Long productId, @Param("productOptionName") String productOptionName);
 
 
     @Modifying
     @Query(value = "UPDATE PRODUCT_OPTION_INFO p SET p.PRODUCT_QUANTITY = :productQuantity WHERE p.PRODUCT_ID = :productId AND p.PRODUCT_OPTION_ID = :productOptionId", nativeQuery = true)
-    ProductQuantityInfraEntity updateByProductId(@Param("productId") long productid, @Param("productOptionId") long productOptionId, @Param("productQuantity") int productQuantity);
+    ProductQuantityInfraEntity updateByProductId(@Param("productId") long productid, @Param("productOptionId") long productOptionId, @Param("productQuantity") long productQuantity);
 }
